SystemUI: Fix blinking when reported battery level is 0
Canvas::drawArc has a bug where, if the sweepAngle is 0,
depending on the startAngle it will draw an empty or full
circle. In the rare cases where the battery reports 0%
this would cause the circular meter to blink furiously
as it animates through several values of startAngle.

As a workaround, let's avoid drawing the percentage bar
if the reported battery level is zero.

Change-Id: I4c1c767f3616c9d2f8a60df63d64ec1f4e266583