Re-use orientation from windows if screen is freezing by apps
This patch prevents the app window being re-created repeatedly. With
the original logic, if an app which is show-when-locked, which has a
different orientation from the keyguard's, and which doesn't handle
config changes, is launched when the screen is off, then it would be
re-created repeatedly.

The original logic prevents this issue when the display is frozen.
However, if the screen is off, the display won't be frozen. So this
change uses additional variables to detect if there are any apps re-
creating their windows.

Change-Id: I73d2e384c4ceeeb38d04d04f0665f83cc6fba7ae
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>